Added fragment_add_giro.xml file

+++ app/src/main/java/dev/bozlak/followcurrentinventorydifference/views/adapters/EventAdapter.java
@@ -5,6 +5,7 @@ import android.view.ViewGroup;
 import android.widget.Toast;
 
 import androidx.annotation.NonNull;
+import androidx.appcompat.app.AlertDialog;
 import androidx.recyclerview.widget.RecyclerView;
 
 import java.sql.Date;
@@ -62,12 +63,25 @@ public class EventAdapter  extends RecyclerView.Adapter<EventAdapter.EventViewHo
         );
 
         holder.binding.btnDeleteInRecyclerRowEvent.setOnClickListener(v -> {
-            int eventId = this.events.get(position).getEventId();
-            if(this.deleteEventGivenEventId(eventId, position)){
-                Toast.makeText(v.getContext(), "Olay silindi!", Toast.LENGTH_SHORT).show();
-            } else {
-                Toast.makeText(v.getContext(), "Olay silme başarısız!!!", Toast.LENGTH_SHORT).show();
-            }
+            AlertDialog.Builder builder = new AlertDialog.Builder(v.getContext());
+            builder.setTitle("Olay Silme İşlemi");
+            String productName = this.events.get(position).getProductName();
+            double eventAmount = this.events.get(position).getEventAmount();
+            String eventDateAndTime = dateString;
+            String message = "Ürün adı : " + productName
+                    + "\nMiktar : " + eventAmount
+                    + "\nTarih : " + eventDateAndTime
+                    + "\n\nBu olayı silmek istediğinize emin misiniz?";
+            builder.setMessage(message);
+            builder.setPositiveButton("Evet", (dialog, which) -> {
+                int eventId = this.events.get(position).getEventId();
+                if(this.deleteEventGivenEventId(eventId, position)){
+                    Toast.makeText(v.getContext(), "Olay silindi!", Toast.LENGTH_SHORT).show();
+                } else {
+                    Toast.makeText(v.getContext(), "Olay silme başarısız!!!", Toast.LENGTH_SHORT).show();
+                }
+            });
+            builder.create().show();
         });
     }
